Unified the two projects into one
