+++ .gitignore
@@ -1,23 +1,14 @@
-*.orig
-*.pyc
-*.swp
-.env
-
-/.cargo_home/
-/.idea/
-/.vs/
-/.vscode/
-!/.vscode/settings.json
-
-# MacOS generated files
 .DS_Store
-.DS_Store?
-
-/ext/websocket/autobahn/reports
+node_modules
+/build
+/.svelte-kit
+/package
+.env
+.env.*
+!.env.example
+vite.config.js.timestamp-*
+vite.config.ts.timestamp-*
 
 # For download and handling
-/data
-
-node_modules/
-
-/dist
+data/*
+!data/chatbot-configuration.txt
