+++ .npmrc
@@ -1,0 +1,2 @@
+engine-strict=true
+resolution-mode=highest
