--- client/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "client",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
-    "check": "svelte-check --tsconfig ./tsconfig.json"
-  },
-  "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^2.0.0",
-    "@tsconfig/svelte": "^3.0.0",
-    "@types/marked": "^4.0.8",
-    "@types/prismjs": "^1.26.0",
-    "autoprefixer": "^10.4.14",
-    "daisyui": "^2.51.6",
-    "postcss": "^8.4.23",
-    "svelte": "^3.54.0",
-    "svelte-check": "^2.10.0",
-    "tailwindcss": "^3.3.2",
-    "tslib": "^2.4.1",
-    "typescript": "^4.9.3",
-    "vite": "^4.0.0"
-  },
-  "dependencies": {
-    "axios": "^1.3.6",
-    "marked": "^4.3.0",
-    "prismjs": "^1.29.0"
-  }
-}
