--- client/postcss.config.cjs
@@ -1,6 +1,0 @@
-module.exports = {
-    plugins: {
-        tailwindcss: {},
-        autoprefixer: {},
-    },
-}
