--- client/svelte.config.js
@@ -1,9 +1,0 @@
-import preprocess from 'svelte-preprocess'
-
-export default {
-  preprocess: [
-    preprocess({
-      postcss: true,
-    }),
-  ],
-}
