--- client/tailwind.config.cjs
@@ -1,11 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-    content: ['./src/**/*.{js,svelte,ts}', './index.html'],
-    theme: {
-        extend: {},
-    },
-    plugins: [require('daisyui')],
-    daisyui: {
-        themes: ['synthwave'],
-    }
-}
