+++ src/routes/+layout.svelte
@@ -1,0 +1,5 @@
+<script>
+	import '../app.postcss';
+</script>
+
+<slot />
